requirements.txt
SQLAlchemy~=2.0.40
fastapi~=0.115.12
pydantic~=2.11.4
uvicorn~=0.34.2
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.9
alembic~=1.13.1
python-dotenv~=1.0.1
email-validator~=2.1.0.post1
pytest~=8.0.0
httpx~=0.27.0
Pillow~=10.2.0
requests~=2.31.0
python-magic~=0.4.27